fix: adjusted max width of open data separator bar
